+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project (attempts to) adhere to [Semantic Versioning](http://semver.org/).
 
+## Uncommitted
+
 ## [1.0.0] - 2017-07-14
 ### Added
 - bonus --check_dependencies now exists